add option to ignore blended sources
A major contributer to the time spent in HSM is iterating over all
the pixels of large, blended sources.  Those measurements aren't
meaningful anyway (due to the blends), so ignore them.  This requires
the user to specify a field for the number of deblend children so
we can test whether the source is blended.