find_ellipmom_1: only evaluate exp for small radii
exp() can account for a large fraction of time spent in HSM,
but at large radii it is almost zero.  Chose to cut it off at
(x-x0)/sigma = 10.  This saves wasting time on pixels that have
no bearing on the result.

This is a compile-time constant (#define) because I didn't want
to go to the trouble of changing the API to get it down to the
bowels, and I didn't want to make it a static variable.