fix ImageConverter::getImageView
I had assumed that GalSim indexed into the array with x + nx*y, but
it actually indexes with x-x0 + nx*(y-y0), so we need to advance
into the array for GalSim to read it correctly.

Also, the GalSim standard is that Bounds are inclusive (in particular,
the upper bound, which is exclusive in LSST).